src/requirements.txt
pandas 
streamlit
scikit-learn
mljar-supervised
